extracted title parsing to separate file
